[Silver V] Title: 김인천씨의 식료품가게 (Small), Time: 228 ms, Memory: 19924 KB -BaekjoonHub

+++ 백준/Silver/12033. 김인천씨의 식료품가게 （Small）/김인천씨의 식료품가게 （Small）.java
@@ -14,7 +14,7 @@ public class Main {
             long[] arr = new long[2 * N];
 
             for (int j = 0; j < 2 * N; j++) {
-                arr[j] = sc.nextInt();
+                arr[j] = sc.nextLong();
             }
 
             Arrays.sort(arr);
